requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
djongo==1.3.7
Faker==33.0.0
pymongo==3.11.4
PyMySQL==1.1.1
python-decouple==3.8
